requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.1.0
openai==1.30.1
numpy==1.26.4
scikit-learn==1.4.1.post1
python-multipart==0.0.6
aiofiles==23.2.1
kaggle==1.5.13
requests==2.31.0
pyyaml==6.0
